requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
openai
unstructured
unstructured[local-inference]
pinecone-client
tiktoken
pypdf
pandas
numpy
python-dotenv
langchain-community